+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Added: `--log-rotate` option to control how often the log file rotates.
+* Added: `--log-retain` option to control how many old log files are retained.
+
 ## v1.3.0 (2023-10-01)
 
 * Fixed: The path to the Shawl executable was not quoted when it contained spaces.